refact(about, schedule, detail, profile): separate view from view controller
